Cleaning up temp changes and routes
