+++ app.js
@@ -23,6 +23,7 @@ const routes = require('./app/routes').router;
 const documentationRoutes = require('./docs/documentation_routes');
 const utils = require('./lib/utils.js');
 
+const openingTimesRoutes = require('./app/routes/openingTimes').router;
 const pocRoutes = require('./app/pocRoutes').router;
 const localStorage = require('./app/routes').localStorage;
 
@@ -192,6 +193,7 @@ if (onlyDocumentation == 'true') {
 // Use custom application routes
 app.use('/', routes);
 app.use('/proof-of-concept', pocRoutes);
+app.use('/editor/opening-times', openingTimesRoutes);
 
 // Automatically route pages
 app.get(/^([^.]+)$/, function(req, res, next) {
