+++ app/routes.js
@@ -10,14 +10,9 @@ const app = require('../app');
 // Create local storage for opening times
 const localStorage = new LocalStorage('./scratch');
 
-moment.locale('en-GB');
-// Documentation router
-router.get('/', function (req, res) {
-  // Save default opening times to local storage
-  localStorage.setItem('localOpeningTimes', JSON.stringify(emptyOpeningTimes));
-  res.render('index');
-});
+let recentChangeMade = false;
 
+moment.locale('en-GB');
 // Branching - Leave a review
 
 router.post('/rating-review', function (req, res) {
@@ -479,290 +474,6 @@ router.post('/facilities', function (req, res) {
   res.redirect('/editor/manage-profile');
 });
 
-const days = {
-  MONDAY: {
-    key: 'monday',
-    display: 'Monday',
-  },
-  TUESDAY: {
-    key: 'tuesday',
-    display: 'Tuesday',
-  },
-  WEDNESDAY: {
-    key: 'wednesday',
-    display: 'Wednesday',
-  },
-  THURSDAY: {
-    key: 'thursday',
-    display: 'Thursday',
-  },
-  FRIDAY: {
-    key: 'friday',
-    display: 'Friday',
-  },
-  SATURDAY: {
-    key: 'saturday',
-    display: 'Saturday',
-  },
-  SUNDAY: {
-    key: 'sunday',
-    display: 'Sunday',
-  },
-};
-
-const getDay = (req) => days[req.params.day.toUpperCase()];
-let recentChangeMade = false;
-let changeMade = '';
-
-// Create default opening times
-const emptyOpeningTimes = Object.keys(days).map((key) => ({
-  name: days[key].display,
-  times: [],
-}));
-
-// Save default opening times to local storage
-localStorage.setItem('localOpeningTimes', JSON.stringify(emptyOpeningTimes));
-localStorage.setItem('lastUpdatedOn', '23 March 2019');
-
-router.get('/editor/opening-times/days', (_, res) => {
-  res.render('editor/opening-times/days', {
-    openingTimes: JSON.parse(localStorage.getItem('localOpeningTimes')),
-    lastUpdatedOn: localStorage.getItem('lastUpdatedOn'),
-  });
-});
-
-router.get('/editor/opening-times/confirm-opening-times', (_, res) => {
-  res.render('editor/opening-times/confirm-opening-times', {
-    openingTimes: JSON.parse(localStorage.getItem('localOpeningTimes')),
-  });
-});
-
-router.get('/editor/opening-times/days/:day', (req, res) => {
-  const dayObj = getDay(req);
-  res.render('editor/opening-times/day', { dayObj });
-});
-
-router.post('/editor/opening-times/days/:day/set', (req, res) => {
-  // Existing opening times from localstorage
-  const localOpeningTimes = JSON.parse(
-    localStorage.getItem('localOpeningTimes')
-  );
-  const dayObj = getDay(req);
-  const currentDay = dayObj.display;
-  const getOpeningTimes = localOpeningTimes.filter((times) => {
-    return times.name == currentDay;
-  });
-  const currentDayOpeningTimes = getOpeningTimes[0].times;
-
-  if (req.body.open === 'yes') {
-    res.render('editor/opening-times/set', { dayObj, currentDayOpeningTimes });
-  } else {
-    const newOpeningTimes = localOpeningTimes.map(({ name, times }) => ({
-      name,
-      times: dayObj.display === name ? [] : times,
-    }));
-    console.log(newOpeningTimes);
-
-    const dateNow = '31 December 2019';
-    // Set localstorage times to new times
-    localStorage.setItem('localOpeningTimes', JSON.stringify(newOpeningTimes));
-    localStorage.setItem('lastUpdatedOn', dateNow);
-    res.redirect('/editor/opening-times/days');
-  }
-});
-
-router.post('/editor/opening-times/days/:day/copy', (req, res) => {
-  const dayObj = getDay(req);
-  const daysToDisplay = { ...days };
-  delete daysToDisplay[dayObj.key.toUpperCase()];
-  res.render('editor/opening-times/copy', {
-    dayObj,
-    daysToDisplay,
-    times: req.body,
-  });
-});
-
-router.post('/editor/opening-times/days', (req, res) => {
-  // Existing opening times from localstorage
-  const localOpeningTimes = JSON.parse(
-    localStorage.getItem('localOpeningTimes')
-  );
-
-  // Get an array of opening times from the request body
-  const newTimes = Object.keys(req.body)
-    .filter((key) => !days[key.toUpperCase()])
-    .map((timeKey) => req.body[timeKey])
-    .filter(Boolean);
-
-  // Build opening times object by adding times to all days sent in the request body
-  const newOpeningTimes = localOpeningTimes.map(({ name, times }) => ({
-    name,
-    times: req.body[name.toLowerCase()] ? newTimes : times,
-  }));
-
-  // Set localstorage times to new times
-  localStorage.setItem('localOpeningTimes', JSON.stringify(newOpeningTimes));
-  const dateNow = moment().format('DD MMMM YYYY');
-  localStorage.setItem('lastUpdatedOn', dateNow);
-  // Redirect to page which will display times from new localstorage
-  res.redirect('/editor/opening-times/confirm-opening-times');
-});
-
-router.post('/editor/opening-times/bank-holiday/day', (req, res) => {
-  if (req.body.open === 'yes') {
-    res.redirect('/editor/opening-times/bank-holiday/opening-time');
-  } else {
-    res.redirect('/editor/opening-times/bank-holiday/days');
-  }
-});
-
-router.post('/editor/opening-times/bank-holiday/opening-time', (req, res) => {
-  let firstTime;
-  let secondTime;
-  let thirdTime;
-  if (req.body.bankHolOpenTime1) {
-    firstTime = `${req.body.bankHolOpenTime1} to ${req.body.bankHolCloseTime1}`;
-  }
-  if (req.body.bankHolOpenTime2) {
-    secondTime = `${req.body.bankHolOpenTime2} to ${req.body.bankHolCloseTime2}`;
-  }
-  if (req.body.bankHolOpenTime3) {
-    thirdTime = `${req.body.bankHolOpenTime3} to ${req.body.bankHolCloseTime3}`;
-  }
-  res.render('editor/opening-times/bank-holiday/days', {
-    firstTime,
-    secondTime,
-    thirdTime,
-  });
-});
-
-localStorage.setItem('tempChanges', JSON.stringify([]));
-
-router.get(
-  '/editor/opening-times/temporary-changes/temporary-changes',
-  (req, res) => {
-    const tempChanges = JSON.parse(localStorage.getItem('tempChanges'));
-    if (tempChanges.length === 0) {
-      res.redirect(
-        '/editor/opening-times/temporary-changes/temporary-changes-date'
-      );
-    } else {
-      res.render('editor/opening-times/temporary-changes/temporary-changes', {
-        tempChanges,
-      });
-    }
-  }
-);
-
-// router.post(
-//   '/editor/opening-times/temporary-changes/temporary-changes-date',
-//   (req, res) => {
-//     let tempDate = `${req.body.tempChangeYear}-${req.body.tempChangeMonth}-${req.body.tempChangeDay}`;
-//     tempDate = moment(tempDate).format('DD MMMM YYYY');
-//     let tempDay = req.body.tempChangeDay
-//     res.redirect(
-//       `/editor/opening-times/temporary-changes/temporary-changes-open?date=${tempDate}`
-//     );
-//   }
-// );
-
-router.post(
-  '/editor/opening-times/temporary-changes/temporary-changes-date',
-  (req, res) => {
-    res.redirect('/editor/opening-times/temporary-changes/temporary-changes-open');
-  }
-);
-
-router.post(
-  '/editor/opening-times/temporary-changes/temporary-change-open', (req, res) => {
-    if (req.body.open === 'yes') {
-      res.redirect(`/editor/opening-times/temporary-changes/temporary-changes-time`);
-    } else {
-      res.redirect('/editor/opening-times/temporary-changes/temporary-changes-range-question')
-    }
-  }
-)
-
-router.post(
-  '/editor/opening-times/temporary-changes/temporary-change-open-one-day', (req, res) => {
-    if (req.body.open === 'yes') {
-      res.redirect(`/editor/opening-times/temporary-changes/temporary-changes-time`);
-    } else {
-      res.redirect('/editor/opening-times/temporary-changes/temporary-changes-done-one-day')
-    }
-  }
-)
-
-// Old fancy way
-// router.post(
-//   '/editor/opening-times/temporary-changes/temporary-change-open',
-//   (req, res) => {
-//     const tempChanges = JSON.parse(localStorage.getItem('tempChanges'));
-//     if (req.body.open === 'yes') {
-//       res.render(
-//         `editor/opening-times/temporary-changes/temporary-changes-time`,
-//         { temporaryDate: req.body.tempDateHidden }
-//       );
-//     } else {
-//       let tempChange = {
-//         date: req.body.tempDateHidden,
-//         time: 'Closed'
-//       };
-//       const newChanges = [...tempChanges, tempChange];
-//       localStorage.setItem('tempChanges', JSON.stringify(newChanges));
-//       res.render('editor/opening-times/temporary-changes/temporary-changes', {
-//         tempChanges: newChanges
-//       });
-//     }
-//   }
-// );
-
-// Old fancy way
-// router.post(
-//   '/editor/opening-times/temporary-changes/temporary-changes-time',
-//   (req, res) => {
-//     const tempChanges = JSON.parse(localStorage.getItem('tempChanges'));
-//     function getTime() {
-//       let time = [];
-//       if (req.body.tempOpenTime1) {
-//         time.push(`${req.body.tempOpenTime1} to ${req.body.tempCloseTime1}`);
-//       }
-//       if (req.body.tempOpenTime2) {
-//         time.push(`${req.body.tempOpenTime2} to ${req.body.tempCloseTime2}`);
-//       }
-//       if (req.body.tempOpenTime3) {
-//         time.push(`${req.body.tempOpenTime3} to ${req.body.tempCloseTime3}`);
-//       }
-//       return time;
-//     }
-//     let tempChange = {
-//       date: req.body.tempDateHidden,
-//       time: getTime()
-//     };
-//     const newChanges = [...tempChanges, tempChange];
-//     localStorage.setItem('tempChanges', JSON.stringify(newChanges));
-//     res.render('editor/opening-times/temporary-changes/temporary-changes', {
-//       tempChanges: newChanges
-//     });
-//   }
-// );
-
-router.post('/editor/opening-times/temporary-changes/temporary-changes-range-question', (req, res) => {
-  if (req.body.tempChangeMultiDay === 'yes') {
-    res.redirect(`/editor/opening-times/temporary-changes/temporary-changes-done-multiple-days`);
-  } else {
-    res.redirect('/editor/opening-times/temporary-changes/temporary-changes-done-one-day');
-  }
-});
-
-router.post('/editor/opening-times/temporary-changes/temporary-changes-range-end', (req, res) => {
-  res.redirect('/editor/opening-times/temporary-changes/temporary-changes-open');
-});
-
-router.post('/editor/opening-times/temporary-changes/temporary-changes-done-multiple-days', (req, res) => {
-  res.redirect('/editor/opening-times/temporary-changes/temporary-changes-done');
-});
-
 router.get('/profiles', (req, res) => {
   let currentPharmacies = pharmacies.slice(0, 15);
   res.render('profiles/index', { currentPharmacies, pharmacies });
@@ -892,4 +603,5 @@ router.get('/profiles/sort-city', (req, res) => {
 module.exports = {
   router: router,
   localStorage: localStorage,
+  recentChangeMade
 };
