--- app/views/editor/opening-times/temporary-changes/temporary-changes-open-one-day.html
@@ -1,113 +1,0 @@
-<!-- Extends the layout from /docs/views/layouts/layout.html -->
-{% extends "layout.html" %}
-
-<!-- Adds a custom class to the <main> wrapper -->
-<!-- nhsuk-main-wrapper--s reduces top and bottom padding of the page -->
-{% set mainClasses = 'nhsuk-main-wrapper--s' %}
-
-<!-- Set the page title inside this block -->
-{% block pageTitle %}
-  Profile Manager
-{% endblock %}
-
-<!-- Header goes inside this block -->
-{% block header %}
-  {{ header({
-    "transactionalService": {
-        "name": "Profile Manager",
-        "href": "/editor"
-      },
-      "showNav": "false",
-      "showSearch": "false"
-    })
-  }}
-{% endblock %}
-
-<!-- Breadcrumb goes inside this block -->
-{% block beforeContent %}
-<nav class="nhsuk-breadcrumb" aria-label="Breadcrumb">
-  <div class="nhsuk-width-container">
-    <ol class="nhsuk-breadcrumb__list">
-      <li class="nhsuk-breadcrumb__item"><a class="nhsuk-breadcrumb__link" href="/editor">Home</a></li>
-      <li class="nhsuk-breadcrumb__item"><a class="nhsuk-breadcrumb__link" href="/editor/manage-profile">
-        {% set displayname = 'alias0' %}
-        {% if data["displayname"] == alias0 %}
-        Edit Practice 1
-        {%else%}
-        Edit {{data["displayname"]}}
-        {%endif%}</a></li>
-      <li class="nhsuk-breadcrumb__item"><a class="nhsuk-breadcrumb__link" href="/editor/opening-times/opening-times-start">Opening times</a></li>
-      <!--li class="nhsuk-breadcrumb__item"><a class="nhsuk-breadcrumb__link" href="/editor/opening-times/temporary-changes/temporary-changes-date">Temporary changes</a></li-->
-    </ol>
-    <p class="nhsuk-breadcrumb__back"><a class="nhsuk-breadcrumb__backlink" href="/editor/manage-profile">
-      {% set displayname = 'alias0' %}
-      {% if data["displayname"] == alias0 %}
-      Back to Edit Practice 1
-      {%else%}
-      Back to Edit {{data["displayname"]}}
-      {%endif%}
-    </a></p>
-  </div>
-</nav>
-{% endblock %}
-
-{% block content %}
-<div class="nhsuk-grid-row">
-  <div class="nhsuk-grid-column-two-thirds">
-    <form method="post" action="/editor/opening-times/temporary-changes/temporary-change-open-one-day">
-      <div class="nhsuk-form-group">
-        <fieldset class="nhsuk-fieldset nhsuk-u-margin-bottom-5">
-          <legend class="nhsuk-fieldset__legend">
-            <h1>{% set displayname = 'alias0' %}
-              {% if data["displayname"] == alias0 %}
-              Is Practice 1 open on 1 June 2021?
-              {%else%}
-              Is {{data["displayname"]}} open on 1 June 2021?
-              {%endif%}</h1>
-          </legend>
-          <input type="text" hidden id="tempDateHidden" name="tempDateHidden" value="{{temporaryDate}}" />
-          <div class="nhsuk-radios nhsuk-radios--inline">
-            <div class="nhsuk-radios__item">
-              <input id="open" class="nhsuk-radios__input" type="radio" value="yes" name="open">
-              <label for="open" class="nhsuk-label nhsuk-radios__label">Yes</label>
-            </div>
-            <div class="nhsuk-radios__item">
-              <input id="closed" class="nhsuk-radios__input" type="radio" value="no" name="open">
-              <label for="closed" class="nhsuk-label nhsuk-radios__label">No</label>
-            </div>
-          </div>
-        </fieldset>
-        <button type="submit" class="nhsuk-button">Continue</button>
-      </div>
-    </form>
-
-    {{ backLink({
-    "href": "/editor/opening-times/temporary-changes/temporary-changes-date",
-    "text": "Back"
-    }) }}
-
-  </div>
-</div>
-
-{% endblock %}
-
-<!-- Footer goes inside this block -->
-
-{% block footer %}
-  {{ footer({
-    "links": [
-      {
-        "URL": "#",
-        "label": "Contact us"
-      },
-      {
-        "URL": "#",
-        "label": "Our policies"
-      },
-      {
-        "URL": "/examples/passing-data/clear-data",
-        "label": "Clear data"
-      }
-    ]
-  })}}
-{% endblock %}
