--- app/views/editor/opening-times/temporary-changes/temporary-changes-range-end.html
@@ -1,123 +1,0 @@
-<!-- Extends the layout from /docs/views/layouts/layout.html -->
-{% extends "layout.html" %}
-
-<!-- Adds a custom class to the <main> wrapper -->
-<!-- nhsuk-main-wrapper--s reduces top and bottom padding of the page -->
-{% set mainClasses = 'nhsuk-main-wrapper--s' %}
-
-<!-- Set the page title inside this block -->
-{% block pageTitle %}
-  Profile Manager
-{% endblock %}
-
-<!-- Header goes inside this block -->
-{% block header %}
-  {{ header({
-    "transactionalService": {
-        "name": "Profile Manager",
-        "href": "/editor"
-      },
-      "showNav": "false",
-      "showSearch": "false"
-    })
-  }}
-{% endblock %}
-
-<!-- Breadcrumb goes inside this block -->
-{% block beforeContent %}
-<nav class="nhsuk-breadcrumb" aria-label="Breadcrumb">
-  <div class="nhsuk-width-container">
-    <ol class="nhsuk-breadcrumb__list">
-      <li class="nhsuk-breadcrumb__item"><a class="nhsuk-breadcrumb__link" href="/editor">Home</a></li>
-      <li class="nhsuk-breadcrumb__item"><a class="nhsuk-breadcrumb__link" href="/editor/manage-profile">
-        {% set displayname = 'alias0' %}
-        {% if data["displayname"] == alias0 %}
-        Edit Practice 1
-        {%else%}
-        Edit {{data["displayname"]}}
-        {%endif%}</a></li>
-      <li class="nhsuk-breadcrumb__item"><a class="nhsuk-breadcrumb__link" href="/editor/opening-times/opening-times-start">Opening times</a></li>
-    </ol>
-    <p class="nhsuk-breadcrumb__back"><a class="nhsuk-breadcrumb__backlink" href="/editor/manage-profile">
-      {% set displayname = 'alias0' %}
-      {% if data["displayname"] == alias0 %}
-      Back to Edit Practice 1
-      {%else%}
-      Back to Edit {{data["displayname"]}}
-      {%endif%}
-    </a></p>
-  </div>
-</nav>
-{% endblock %}
-
-{% block content %}
-<div class="nhsuk-grid-row">
-  <div class="nhsuk-grid-column-two-thirds">
-    <div class="nhsuk-form-group">
-      <form action="/editor/opening-times/temporary-changes/temporary-changes-range-end" method="POST">
-        <fieldset class="nhsuk-fieldset" aria-describedby="dob-hint" role="group">
-          <legend class="nhsuk-fieldset__legend">
-            <h1>When does the temporary change end?</h1>
-          </legend>
-          <span class="nhsuk-hint">
-            For example, 31 12 2020 
-          </span>
-          <div class="nhsuk-date-input nhsuk-u-margin-bottom-5">
-            <div class="nhsuk-date-input__item">
-              <div class="nhsuk-form-group">
-                <label class="nhsuk-label nhsuk-date-input__label" for="tempChangeDay">
-                  Day
-                </label>
-                <input class="nhsuk-input nhsuk-date-input__input nhsuk-input--width-2" id="tempChangeDay" name="tempChangeDay" type="number" pattern="[0-9]*">
-              </div>
-            </div>
-            <div class="nhsuk-date-input__item">
-              <div class="nhsuk-form-group">
-                <label class="nhsuk-label nhsuk-date-input__label" for="tempChangeMonth">
-                  Month
-                </label>
-                <input class="nhsuk-input nhsuk-date-input__input nhsuk-input--width-2" id="tempChangeMonth" name="tempChangeMonth" type="number" pattern="[0-9]*">
-              </div>
-            </div>
-            <div class="nhsuk-date-input__item">
-              <div class="nhsuk-form-group">
-                <label class="nhsuk-label nhsuk-date-input__label" for="tempChangeYear">
-                  Year
-                </label>
-                <input class="nhsuk-input nhsuk-date-input__input nhsuk-input--width-3" id="tempChangeYear" name="tempChangeYear" type="number" pattern="[0-9]*">
-              </div>
-            </div>
-          </div>
-        </fieldset>
-        <button type="submit" class="nhsuk-button">Continue</button>
-      </form>
-    </div>
-    {{ backLink({
-    "href": "/editor/opening-times/opening-times-start",
-    "text": "Back"
-    }) }}
-  </div>
-</div>
-
-{% endblock %}
-
-<!-- Footer goes inside this block -->
-
-{% block footer %}
-  {{ footer({
-    "links": [
-      {
-        "URL": "#",
-        "label": "Contact us"
-      },
-      {
-        "URL": "#",
-        "label": "Our policies"
-      },
-      {
-        "URL": "/examples/passing-data/clear-data",
-        "label": "Clear data"
-      }
-    ]
-  })}}
-{% endblock %}
